use mnemonic for keystores
