+++ src/derive-key/index.ts
@@ -1,1 +1,2 @@
 export * from './deriveKeysFromKey';
+export * from './deriveKeysFromMnemonic';
